chore: :truck: move search events filter file location
