+++ lib/page/search/search_events.dart
@@ -14,7 +14,7 @@ import 'package:flutter_boilerplate/event/data/common/event_model.dart';
 import 'package:flutter_boilerplate/event/data/search_event/filter_event_page_model.dart';
 import 'package:flutter_boilerplate/event/data/search_event/search_event_repository.dart';
 import 'package:flutter_boilerplate/page/event/event_detail.dart';
-import 'package:flutter_boilerplate/page/search_events_filter.dart';
+import 'package:flutter_boilerplate/page/search/search_events_filter.dart';
 
 class SearchEvents extends StatefulWidget {
   const SearchEvents({Key? key}) : super(key: key);
